+++ changelog.md
@@ -4,6 +4,8 @@ All notable changes to `ImageResizer` will be documented in this file.
 
 ## VERSION 0.1.2
 - Fixed Using Storage instead of blind path, thus this package now need the usage of storage:link
+- Added command `php artisan image-resizer:clean-cache`
+- Added The previous command is **automatically scheduled every half hour**
 
 ## VERSION 0.1.1
 - Fixed strip prepended slash on source string